Migrates genesis logic from block to vm
