+++ parameters/examples/inclusion.rs
@@ -88,7 +88,7 @@ pub fn sample_assignment<N: Network, A: Aleo<Network = N>>() -> Result<(Assignme
     // Initialize a new caller.
     let caller_private_key = PrivateKey::<N>::new(rng).unwrap();
     // Return the block.
-    let genesis_block = Block::genesis(&vm, &caller_private_key, rng)?;
+    let genesis_block = vm.genesis(&caller_private_key, rng)?;
 
     // Update the VM.
     vm.add_next_block(&genesis_block)?;
