+++ synthesizer/block/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "snarkvm-synthesizer-block"
+version = "0.11.2"
+authors = [ "The Aleo Team <hello@aleo.org>" ]
+description = "A block for a decentralized virtual machine"
+homepage = "https://aleo.org"
+repository = "https://github.com/AleoHQ/snarkVM"
+keywords = [
+  "aleo",
+  "cryptography",
+  "blockchain",
+  "decentralized",
+  "zero-knowledge"
+]
+categories = [
+  "compilers",
+  "cryptography",
+  "mathematics",
+  "wasm",
+  "web-programming"
+]
+include = [ "Cargo.toml", "src", "README.md", "LICENSE.md" ]
+license = "GPL-3.0"
+edition = "2021"
